requirements.txt
numpy==2.4.4
pyarrow==23.0.1
rerun-sdk==0.31.3
